Improve portability by integral division
This is a continuation after the commit 755f31e75a59 ("3.12.0")
which removed the dependency to the float log.

Once we find quartile values in TLSH, we perform a float division to
get q1/q3 and q2/q3 in percentage (and taken the mod 16) but lacks
portability (and has a bug which the committer explains later).

If we have a 32-bit divider, a 64-bit / 32-bit division is not that a
big cost (it's theoretically one instruction in x86 because the DIV
instruction performs a 64-bit / 32-bit division).

It also fixes the problem of arithmetic overflow calculating
unsigned 32-bit (q1*100) and (q2*100) (if they overflow, the ratio
will be an unexpected value).